+++ src/dodal/devices/areadetector/plugins/MJPG.py
@@ -10,10 +10,10 @@ class MJPG(Device):
     filename: Signal = Component(Signal)
     directory: Signal = Component(Signal)
     url: EpicsSignal = Component(EpicsSignal, "JPG_URL_RBV", string=True)
-    x_size_pv: EpicsSignalRO = Component(EpicsSignalRO, "ArraySize1_RBV")
-    y_size_pv: EpicsSignalRO = Component(EpicsSignalRO, "ArraySize2_RBV")
+    x_size: EpicsSignalRO = Component(EpicsSignalRO, "ArraySize1_RBV")
+    y_size: EpicsSignalRO = Component(EpicsSignalRO, "ArraySize2_RBV")
     input_rbpv: EpicsSignalRO = Component(EpicsSignalRO, "NDArrayPort_RBV")
-    input_pv: EpicsSignal = Component(EpicsSignal, "NDArrayPort")
+    input_plugin_pv: EpicsSignal = Component(EpicsSignal, "NDArrayPort")
     KICKOFF_TIMEOUT: float = 10.0
 
     def trigger(self):
