+++ src/dodal/devices/areadetector/plugins/MXSC.py
@@ -1,17 +1,4 @@
-from dodal.devices.oav.grid_overlay import SnapshotWithGrid
-from dodal.devices.oav.oav_detector import ZoomController
-from ophyd import ADComponent as ADC
-from ophyd import (
-    AreaDetector,
-    CamBase,
-    Component,
-    Device,
-    EpicsSignal,
-    HDF5Plugin,
-    OverlayPlugin,
-    ProcessPlugin,
-    ROIPlugin,
-)
+from ophyd import Component, Device, EpicsSignal
 
 
 class MXSC(Device):
@@ -45,15 +32,3 @@ class MXSC(Device):
     draw_edges: EpicsSignal = Component(EpicsSignal, "DrawEdges")
     waveform_size_x: EpicsSignal = Component(EpicsSignal, "ArraySize1_RBV")
     waveform_size_y: EpicsSignal = Component(EpicsSignal, "ArraySize2_RBV")
-
-
-class OAV(AreaDetector):
-    cam: CamBase = ADC(CamBase, "-DI-OAV-01:CAM:")
-    roi: ADC = ADC(ROIPlugin, "-DI-OAV-01:ROI:")
-    proc: ADC = ADC(ProcessPlugin, "-DI-OAV-01:PROC:")
-    over: ADC = ADC(OverlayPlugin, "-DI-OAV-01:OVER:")
-    tiff: ADC = ADC(OverlayPlugin, "-DI-OAV-01:TIFF:")
-    hdf5: ADC = ADC(HDF5Plugin, "-DI-OAV-01:HDF5:")
-    snapshot: SnapshotWithGrid = Component(SnapshotWithGrid, "-DI-OAV-01:MJPG:")
-    mxsc: MXSC = ADC(MXSC, "-DI-OAV-01:MXSC:")
-    zoom_controller: ZoomController = ADC(ZoomController, "-EA-OAV-01:FZOOM:")
