+++ src/dodal/devices/oav/oav_detector.py
@@ -1,21 +1,18 @@
-from enum import IntEnum
+from ophyd import ADComponent as ADC
+from ophyd import (
+    AreaDetector,
+    CamBase,
+    Component,
+    Device,
+    EpicsSignal,
+    HDF5Plugin,
+    OverlayPlugin,
+    ProcessPlugin,
+    ROIPlugin,
+)
 
-from ophyd import Component, Device, EpicsSignal
-
-
-class ColorMode(IntEnum):
-    """
-    Enum to store the various color modes of the camera. We use RGB1.
-    """
-
-    MONO = 0
-    BAYER = 1
-    RGB1 = 2
-    RGB2 = 3
-    RGB3 = 4
-    YUV444 = 5
-    YUV422 = 6
-    YUV421 = 7
+from dodal.devices.areadetector.plugins.MXSC import MXSC
+from dodal.devices.oav.grid_overlay import SnapshotWithGrid
 
 
 class ZoomController(Device):
@@ -49,13 +46,13 @@ class ZoomController(Device):
         ]
 
 
-class EdgeOutputArrayImageType(IntEnum):
-    """
-    Enum to store the types of image to tweak the output array. We use Original.
-    """
-
-    ORIGINAL = 0
-    GREYSCALE = 1
-    PREPROCESSED = 2
-    CANNY_EDGES = 3
-    CLOSED_EDGES = 4
+class OAV(AreaDetector):
+    cam: CamBase = ADC(CamBase, "-DI-OAV-01:CAM:")
+    roi: ADC = ADC(ROIPlugin, "-DI-OAV-01:ROI:")
+    proc: ADC = ADC(ProcessPlugin, "-DI-OAV-01:PROC:")
+    over: ADC = ADC(OverlayPlugin, "-DI-OAV-01:OVER:")
+    tiff: ADC = ADC(OverlayPlugin, "-DI-OAV-01:TIFF:")
+    hdf5: ADC = ADC(HDF5Plugin, "-DI-OAV-01:HDF5:")
+    snapshot: SnapshotWithGrid = Component(SnapshotWithGrid, "-DI-OAV-01:MJPG:")
+    mxsc: MXSC = ADC(MXSC, "-DI-OAV-01:MXSC:")
+    zoom_controller: ZoomController = ADC(ZoomController, "-EA-OAV-01:FZOOM:")
