+++ tests/devices/system_tests/test_oav_system.py
@@ -2,7 +2,7 @@ import bluesky.plan_stubs as bps
 import pytest
 from bluesky import RunEngine
 
-from dodal.devices.oav.oav_detector import OAV, ZoomController
+from dodal.devices.oav.zoom_controller import OAV, ZoomController
 
 TEST_GRID_TOP_LEFT_X = 100
 TEST_GRID_TOP_LEFT_Y = 100
@@ -13,11 +13,11 @@ TEST_GRID_NUM_BOXES_Y = 6
 
 def take_snapshot_with_grid(oav: OAV, snapshot_filename, snapshot_directory):
     oav.wait_for_connection()
-    yield from bps.abs_set(oav.snapshot.top_left_x_signal, TEST_GRID_TOP_LEFT_X)
-    yield from bps.abs_set(oav.snapshot.top_left_y_signal, TEST_GRID_TOP_LEFT_Y)
-    yield from bps.abs_set(oav.snapshot.box_width_signal, TEST_GRID_BOX_WIDTH)
-    yield from bps.abs_set(oav.snapshot.num_boxes_x_signal, TEST_GRID_NUM_BOXES_X)
-    yield from bps.abs_set(oav.snapshot.num_boxes_y_signal, TEST_GRID_NUM_BOXES_Y)
+    yield from bps.abs_set(oav.snapshot.top_left_x, TEST_GRID_TOP_LEFT_X)
+    yield from bps.abs_set(oav.snapshot.top_left_y, TEST_GRID_TOP_LEFT_Y)
+    yield from bps.abs_set(oav.snapshot.box_width, TEST_GRID_BOX_WIDTH)
+    yield from bps.abs_set(oav.snapshot.num_boxes_x, TEST_GRID_NUM_BOXES_X)
+    yield from bps.abs_set(oav.snapshot.num_boxes_y, TEST_GRID_NUM_BOXES_Y)
     yield from bps.abs_set(oav.snapshot.filename, snapshot_filename)
     yield from bps.abs_set(oav.snapshot.directory, snapshot_directory)
     yield from bps.trigger(oav.snapshot, wait=True)
